backend/requirements.txt
django
mysqlclient
djangorestframework
djangorestframework-simplejwt
django-cors-headers
openpyxl
pandas
dotenv
minio
chardet
django_celery_beat
channels
channels-redis
daphne
drf-nested-routers